cuong

+++ database/migrations/2024_11_04_085236_create_ma_khuyen_mais_table.php
@@ -14,10 +14,10 @@ return new class extends Migration
         Schema::create('ma_khuyen_mais', function (Blueprint $table) {
             $table->id();
             $table->string('ma_giam_gia');
-            $table->decimal('so_tien_giam', 10, 2);
+            $table->double('so_tien_giam', 10, 2);
             $table->date('ngay_bat_dau');
             $table->date('ngay_ket_thuc');
-            $table->decimal('gia_tri_toi_thieu', 10, 2)->default(0);
+            $table->double('gia_tri_toi_thieu', 10, 2)->default(0);
             $table->integer('trang_thai');  // 1. Giảm giá đơn hàng, 2. giảm giá vận chuyển
             $table->timestamps();
         });
